+++ src/parlant/adapters/nlp/cerebras_service.py
@@ -41,7 +41,7 @@ from parlant.core.nlp.generation_info import GenerationInfo, UsageInfo
 from parlant.core.loggers import Logger
 from parlant.core.nlp.moderation import ModerationService, NoModeration
 from parlant.core.nlp.policies import policy, retry
-from parlant.core.nlp.service import NLPService
+from parlant.core.nlp.service import EmbedderHints, NLPService, SchematicGeneratorHints
 from parlant.core.nlp.tokenization import EstimatingTokenizer
 
 
@@ -244,11 +244,13 @@ Please set CEREBRAS_API_KEY in your environment before running Parlant.
         self._logger.info("Initialized CerebrasService")
 
     @override
-    async def get_schematic_generator(self, t: type[T]) -> CerebrasSchematicGenerator[T]:
+    async def get_schematic_generator(
+        self, t: type[T], hints: SchematicGeneratorHints = {}
+    ) -> CerebrasSchematicGenerator[T]:
         return Llama3_3_70B[t](self._logger, self._meter)  # type: ignore
 
     @override
-    async def get_embedder(self) -> Embedder:
+    async def get_embedder(self, hints: EmbedderHints = {}) -> Embedder:
         return JinaAIEmbedder(self._logger, self._meter)
 
     @override