+++ src/parlant/adapters/nlp/deepseek_service.py
@@ -39,7 +39,7 @@ from parlant.core.loggers import Logger
 from parlant.core.meter import Meter
 from parlant.core.nlp.policies import policy, retry
 from parlant.core.nlp.tokenization import EstimatingTokenizer
-from parlant.core.nlp.service import NLPService
+from parlant.core.nlp.service import EmbedderHints, NLPService, SchematicGeneratorHints
 from parlant.core.nlp.embedding import Embedder
 from parlant.core.nlp.generation import (
     T,
@@ -228,11 +228,13 @@ Please set DEEPSEEK_API_KEY in your environment before running Parlant.
         self._logger.info("Initialized DeepSeekService")
 
     @override
-    async def get_schematic_generator(self, t: type[T]) -> DeepSeekSchematicGenerator[T]:
+    async def get_schematic_generator(
+        self, t: type[T], hints: SchematicGeneratorHints = {}
+    ) -> DeepSeekSchematicGenerator[T]:
         return DeepSeek_Chat[t](self._logger, self._meter)  # type: ignore
 
     @override
-    async def get_embedder(self) -> Embedder:
+    async def get_embedder(self, hints: EmbedderHints = {}) -> Embedder:
         return JinaAIEmbedder(self._logger, self._meter)
 
     @override