+++ src/parlant/adapters/nlp/glm_service.py
@@ -38,7 +38,7 @@ from parlant.core.loggers import Logger
 from parlant.core.meter import Meter
 from parlant.core.nlp.policies import policy, retry
 from parlant.core.nlp.tokenization import EstimatingTokenizer
-from parlant.core.nlp.service import NLPService
+from parlant.core.nlp.service import EmbedderHints, NLPService, SchematicGeneratorHints
 from parlant.core.nlp.embedding import BaseEmbedder, Embedder, EmbeddingResult
 from parlant.core.nlp.generation import (
     T,
@@ -311,11 +311,13 @@ Please set GLM_API_KEY in your environment before running Parlant.
         self._logger.info("Initialized GLMService")
 
     @override
-    async def get_schematic_generator(self, t: type[T]) -> GLMSchematicGenerator[T]:
+    async def get_schematic_generator(
+        self, t: type[T], hints: SchematicGeneratorHints = {}
+    ) -> GLMSchematicGenerator[T]:
         return GLM_4_5[t](self._logger, self._meter)  # type: ignore
 
     @override
-    async def get_embedder(self) -> Embedder:
+    async def get_embedder(self, hints: EmbedderHints = {}) -> Embedder:
         return GMLTextEmbedding_3(logger=self._logger, meter=self._meter)
 
     @override