+++ src/parlant/adapters/nlp/snowflake_cortex_service.py
@@ -32,7 +32,7 @@ from parlant.core.loggers import Logger
 from parlant.core.meter import Meter
 from parlant.core.nlp.policies import policy, retry
 from parlant.core.nlp.tokenization import EstimatingTokenizer
-from parlant.core.nlp.service import NLPService
+from parlant.core.nlp.service import EmbedderHints, NLPService, SchematicGeneratorHints
 from parlant.core.nlp.embedding import BaseEmbedder, Embedder, EmbeddingResult
 from parlant.core.nlp.generation import (
     T,
@@ -372,11 +372,13 @@ class SnowflakeCortexService(NLPService):
         )
 
     @override
-    async def get_schematic_generator(self, t: type[T]) -> SchematicGenerator[T]:
+    async def get_schematic_generator(
+        self, t: type[T], hints: SchematicGeneratorHints = {}
+    ) -> SchematicGenerator[T]:
         return CortexSchematicGenerator[t](schema=t, logger=self._logger, meter=self._meter)  # type: ignore[valid-type,misc]
 
     @override
-    async def get_embedder(self) -> Embedder:
+    async def get_embedder(self, hints: EmbedderHints = {}) -> Embedder:
         return CortexEmbedder(logger=self._logger, meter=self._meter)
 
     @override