+++ CHANGELOG.md
@@ -22,6 +22,9 @@ All notable changes to Parlant will be documented here.
 - Add OpenTelemetry integration for Meter, Logger and Tracer
 - Add API for CRUD operations on session metadata
 - Add deferred retrievers
+- Allow controlling perceived performance policy per agent
+- Expose IoC container in server object
+- Support code-based, custom guideline matchers
 
 ## [3.0.3] - 2025-10-23